requirements.txt
torch
torchvision
ultralytics
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python